+++ testapps/ResourceLoaderTest/src/main/java/resourceloader/AppEntry.java
@@ -1,0 +1,12 @@
+package resourceloader;
+
+import act.Act;
+
+@SuppressWarnings("unused")
+public class AppEntry {
+
+    public static void main(String[] args) throws Exception {
+        Act.start();
+    }
+
+}